Splitting abstract test class to create more flexible hierarchy, utility methods separated from specific test scenarios.
