+++ it/src/test/java/com/expediagroup/streamplatform/streamregistry/it/helpers/AbstractTestStage.java
@@ -15,29 +15,13 @@
  */
 package com.expediagroup.streamplatform.streamregistry.it.helpers;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 
-import junit.framework.TestCase;
 
-import com.apollographql.apollo.api.Mutation;
-import com.apollographql.apollo.api.Response;
 
-import org.junit.Before;
 import org.junit.Test;
 
-import com.expediagroup.streamplatform.streamregistry.it.StreamRegistryIT;
 
-public abstract class AbstractTestStage {
-  protected ITestDataFactory factory = new ITestDataFactory(getClass().getSimpleName().toLowerCase());
-
-  protected Client client = StreamRegistryIT.client;
-
-  @Before
-  public final void before() {
-    createRequiredDatastoreState();
-  }
-
+public abstract class AbstractTestStage extends AbstractTest {
   @Test
   public abstract void create();
 
@@ -58,48 +42,4 @@ public abstract class AbstractTestStage {
 
   @Test
   public abstract void queryByRegex();
-
-  public abstract void createRequiredDatastoreState();
-
-  public Response assertMutationSucceeds(Mutation mutation) {
-    try {
-      return client.invoke(mutation);
-    } catch (RuntimeException e) {
-      TestCase.fail("Mutation failed");
-    }
-    return null;
-  }
-
-  public Mutation assertMutationFails(Mutation mutation) {
-    if (mutation.getClass().getSimpleName().toLowerCase().contains("insert")) {
-      assertRequiresObjectIsAbsent(mutation);
-    }
-    if (mutation.getClass().getSimpleName().toLowerCase().contains("update")) {
-      assertRequiresObjectIsPresent(mutation);
-    }
-    return mutation;
-  }
-
-  public void assertRequiresObjectIsAbsent(Mutation m) {
-    try {
-      client.invoke(m);
-      TestCase.fail("Expected a ValidationException");
-    } catch (RuntimeException e) {
-      assertEquals("Can't create because it already exists", e.getMessage());
-    }
-  }
-
-  public void setFactorySuffix(String suffix) {
-    factory = new ITestDataFactory(suffix);
-    createRequiredDatastoreState();
-  }
-
-  public void assertRequiresObjectIsPresent(Mutation m) {
-    try {
-      client.invoke(m);
-      TestCase.fail("Expected a ValidationException");
-    } catch (RuntimeException e) {
-      assertTrue(e.getMessage().startsWith("Can't update"));
-    }
-  }
 }
